# handler/account_handler.go
package handler

import (
	"context"
	"time"

	"github.com/gin-gonic/gin"
	_ "github.com/michaelyusak/go-helper/dto"
	"github.com/michaelyusak/go-helper/helper"
	"github.com/michaelyusak/xyz-kredit-plus/entity"
	"github.com/michaelyusak/xyz-kredit-plus/service"
)

type AccountHandler struct {
	ctxTimeout     time.Duration
	accountService service.AccountService
}

func NewAccountHandler(accountService service.AccountService, ctxTimeout time.Duration) *AccountHandler {
	if ctxTimeout <= 0 {
		ctxTimeout = 30 * time.Second
	}

	return &AccountHandler{
		accountService: accountService,
		ctxTimeout:     ctxTimeout,
	}
}

// Account godoc
// @Summary Register a new account
// @Description Creates a new account
// @Tags accounts
// @Accept  json
// @Produce  json
// @Param request body entity.LoginRegisterReq true "Register request body"
// @Success 200 {object} dto.Response{message=string,data=entity.TokenData} "Success"
// @Router /account/register [post]
func (h *AccountHandler) Register(ctx *gin.Context) {
	ctx.Header("Content-Type", "application/json")

	var newAccount entity.Account

	err := ctx.ShouldBind(&newAccount)
	if err != nil {
		ctx.Error(err)
		return
	}

	ctxWithTimeout, cancel := context.WithTimeout(ctx.Request.Context(), h.ctxTimeout)
	defer cancel()

	token, err := h.accountService.RegisterAccount(ctxWithTimeout, newAccount)
	if err != nil {
		ctx.Error(err)
		return
	}

	helper.ResponseOK(ctx, *token)
}

// Account godoc
// @Summary Login to an account
// @Description Login to an existing account
// @Tags accounts
// @Accept  json
// @Produce  json
// @Param request body entity.LoginRegisterReq true "Register request body"
// @Success 200 {object} dto.Response{message=string,data=entity.TokenData} "Success"
// @Router /account/login [post]
func (h *AccountHandler) Login(ctx *gin.Context) {
	ctx.Header("Content-Type", "application/json")

	var req entity.Account

	err := ctx.ShouldBind(&req)
	if err != nil {
		ctx.Error(err)
		return
	}

	ctxWithTimeout, cancel := context.WithTimeout(ctx.Request.Context(), h.ctxTimeout)
	defer cancel()

	token, err := h.accountService.Login(ctxWithTimeout, req)
	if err != nil {
		ctx.Error(err)
		return
	}

	helper.ResponseOK(ctx, *token)
}